Edit-learning-objectives-dplyr-lesson
I made several changes in the learning objectives to emphasize the functions in the 'dplyr' package. Because this lesson introduces two packages, it's important to make learners understand what are the specific functions in 'dplyr' and 'tidyr' separately. 
p.s. This is one step for my carpentry training check-out. /Wei Yin, Columbia University